Fixed the POJO package location
